+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [#2190](https://github.com/influxdb/influxdb/pull/2190): Implement failover to other data nodes for distributed queries
 - [#2324](https://github.com/influxdb/influxdb/issues/2324): Race in Broker.Close()/Broker.RunContinousQueryProcessing()
 - [#2325](https://github.com/influxdb/influxdb/pull/2325): Cluster open fixes
+- [#2326](https://github.com/influxdb/influxdb/pull/2326): Fix parse error in CREATE CONTINUOUS QUERY
 
 ## v0.9.0-rc25 [2015-04-15]